Move things outside of export_test
The following series information is moved outside of export_test.go
for a couple of reasons:

 1. To make it easier to locate when we bump the series in the future
 2. To allow us to refactor in the future, where by the tests are run
 in the *_test namespace.

We should treat export_test.go sparingly and not overload the reason
it exists.
